add  Decompose monolithic QuestManager.tsx (admin) into smaller subcomponents
