+++ FrontEnd/my-app/components/admin/index.ts
@@ -4,5 +4,15 @@ export { QuestEditor } from './QuestEditor';
 export { QuestForm } from './QuestForm';
 export { QuestEditModal } from './QuestEditModal';
 export { QuestManager } from './QuestManager';
+export { QuestFilters } from './QuestFilters';
+export type { QuestFiltersProps } from './QuestFilters';
+export { QuestBulkActions } from './QuestBulkActions';
+export type { QuestBulkActionsProps, BulkAction } from './QuestBulkActions';
+export { QuestTable } from './QuestTable';
+export type { QuestTableProps } from './QuestTable';
+export { QuestDeleteDialog } from './QuestDeleteDialog';
+export type { QuestDeleteDialogProps } from './QuestDeleteDialog';
+export { useQuestFilters } from './useQuestFilters';
+export type { SortField, SortOrder } from './useQuestFilters';
 export { Notifications } from './Notifications';
 export { Analytics } from './Analytics';
